requirements.txt
discord.py==2.4.2
aiohttp==3.8.1
colorama==0.4.6
